feat(mock-data.ts): move mock-data.ts from sprints to global/mocks directory The mock-data.ts file was moved from the sprints directory to the global/mocks directory. This change was made to improve the organization of the codebase. The mock-data.ts file contains mock data that can be used across multiple tests, not just those in the sprints directory. By moving it to the global/mocks directory, it is now more accessible for use in tests throughout the application.
